# application/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于区块链网红信息追根溯源系统api文档",
        "title": "基于区块链网红信息追根溯源系统api文档",
        "contact": {
            "name": "网红电商组",
            "email": "233_666@qq.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/v1/createRealGoods": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "新建商品",
                "parameters": [
                    {
                        "description": "realGoods",
                        "name": "realGoods",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.RealGoodsRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/createSelling": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "发起销售",
                "parameters": [
                    {
                        "description": "selling",
                        "name": "selling",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.SellingRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/createSellingByBuy": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "买家购买",
                "parameters": [
                    {
                        "description": "sellingByBuy",
                        "name": "sellingByBuy",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.SellingByBuyRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/hello": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "测试输出Hello",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/queryAccountList": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "获取账户信息",
                "parameters": [
                    {
                        "description": "account",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.AccountRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/queryRealGoodsList": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "获取商品信息(空json{}可以查询所有，指定brand可以查询指定品牌旗下的商品)",
                "parameters": [
                    {
                        "description": "realGoodsQuery",
                        "name": "realGoodsQuery",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.RealGoodsQueryRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/querySellingList": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "查询销售(可查询所有，也可根据网红查询)(发起的)",
                "parameters": [
                    {
                        "description": "sellingListQuery",
                        "name": "sellingListQuery",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.SellingListQueryRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/querySellingListByBuyer": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "根据参与销售人、买家(买家AccountId)查询销售(参与的)",
                "parameters": [
                    {
                        "description": "sellingListQueryByBuy",
                        "name": "sellingListQueryByBuy",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.SellingListQueryByBuyRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/updateSelling": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "更新销售状态（买家确认、买卖家取消）Status取值为 完成\"done\"、取消\"cancelled\" 当处于销售中状态，卖家要取消时，buyer为\"\"空",
                "parameters": [
                    {
                        "description": "updateSelling",
                        "name": "updateSelling",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.UpdateSellingRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "app.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "v1.AccountIdBody": {
            "type": "object",
            "properties": {
                "accountId": {
                    "type": "string"
                }
            }
        },
        "v1.AccountRequestBody": {
            "type": "object",
            "properties": {
                "args": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.AccountIdBody"
                    }
                }
            }
        },
        "v1.RealGoodsQueryRequestBody": {
            "type": "object",
            "properties": {
                "Brand": {
                    "description": "卖家(网红)(网红AccountId)",
                    "type": "string"
                }
            }
        },
        "v1.RealGoodsRequestBody": {
            "type": "object",
            "properties": {
                "accountId": {
                    "description": "操作人ID",
                    "type": "string"
                },
                "isBrand": {
                    "description": "是否品牌保证",
                    "type": "number"
                },
                "Brand": {
                    "description": "卖家(网红)(网红AccountId)",
                    "type": "string"
                },
                "isSeller": {
                    "description": "是否卖家认证",
                    "type": "number"
                }
            }
        },
        "v1.SellingByBuyRequestBody": {
            "type": "object",
            "properties": {
                "buyer": {
                    "description": "买家(买家AccountId)",
                    "type": "string"
                },
                "objectOfSale": {
                    "description": "销售对象(正在出售的商品RealGoodsID)",
                    "type": "string"
                },
                "seller": {
                    "description": "发起推荐的网红(卖家AccountId)",
                    "type": "string"
                }
            }
        },
        "v1.SellingListQueryByBuyRequestBody": {
            "type": "object",
            "properties": {
                "buyer": {
                    "description": "买家(买家AccountId)",
                    "type": "string"
                }
            }
        },
        "v1.SellingListQueryRequestBody": {
            "type": "object",
            "properties": {
                "seller": {
                    "description": "发起推荐的网红(卖家AccountId)",
                    "type": "string"
                }
            }
        },
        "v1.SellingRequestBody": {
            "type": "object",
            "properties": {
                "objectOfSale": {
                    "description": "销售对象(正在出售的商品RealGoodsID)",
                    "type": "string"
                },
                "price": {
                    "description": "价格",
                    "type": "number"
                },
                "salePeriod": {
                    "description": "智能合约的有效期(单位为天)",
                    "type": "integer"
                },
                "seller": {
                    "description": "发起推荐的网红(卖家AccountId)",
                    "type": "string"
                }
            }
        },
        "v1.UpdateSellingRequestBody": {
            "type": "object",
            "properties": {
                "buyer": {
                    "description": "买家(买家AccountId)",
                    "type": "string"
                },
                "objectOfSale": {
                    "description": "销售对象(正在出售的商品RealGoodsID)",
                    "type": "string"
                },
                "seller": {
                    "description": "发起推荐的网红(卖家AccountId)",
                    "type": "string"
                },
                "status": {
                    "description": "需要更改的状态",
                    "type": "string"
                }
            }
        }
    }
}