# application/pkg/app/response.go
/**
 * @Author: 网红电商组
 * @Email: 233_666@qq.com
 * @Date: 2020/7/18 12:51 上午
 * @Description: 定义请求返回JSON格式
 */
package app

import (
	"github.com/gin-gonic/gin"
)

type Gin struct {
	C *gin.Context
}

type Response struct {
	Code int         `json:"code"`
	Msg  string      `json:"msg"`
	Data interface{} `json:"data"`
}

func (g *Gin) Response(httpCode int, errMsg string, data interface{}) {
	g.C.JSON(httpCode, Response{
		Code: httpCode,
		Msg:  errMsg,
		Data: data,
	})
	return
}